restructure; __init__ to each module
